+++ src/strategies/SpreadStrategy.cpp
@@ -10,8 +10,8 @@ void SpreadStrategy::onEvent(OrderBook &ob)
     if (ob.bids.empty() || ob.asks.empty())
         return;
 
-    int bestBid = ob.bids.begin()->first;
-    int bestAsk = ob.asks.begin()->first;
+    double bestBid = ob.bids.begin()->first;
+    double bestAsk = ob.asks.begin()->first;
 
     // Kill switch
     if (risk.killSwitch(currentPnL(ob)))
